# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DanceClass {
  id: number;
  name: string;
  style: string;
  description: string;
  instructor: string;
  location: string;
  scheduleDay: string;
  scheduleTime: string;
  price: number;
  pricePeriod: string;
  duration: string;
  ageGroup: string;
  category: string;
  capacity: number;
  spotsAvailable: number;
  colorScheme: string;
  createdAt: string;
}

export interface DanceClassInput {
  name: string;
  style: string;
  description: string;
  instructor: string;
  location: string;
  scheduleDay: string;
  scheduleTime: string;
  price: number;
  pricePeriod: string;
  duration: string;
  ageGroup: string;
  category: string;
  capacity: number;
  spotsAvailable: number;
  colorScheme: string;
}

export interface DanceClassUpdate {
  name?: string;
  style?: string;
  description?: string;
  instructor?: string;
  location?: string;
  scheduleDay?: string;
  scheduleTime?: string;
  price?: number;
  pricePeriod?: string;
  duration?: string;
  ageGroup?: string;
  category?: string;
  capacity?: number;
  spotsAvailable?: number;
  colorScheme?: string;
}

export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  studentType: string;
  createdAt: string;
}

export interface StudentInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  studentType: string;
}

export interface Enrollment {
  id: number;
  studentId: number;
  classId: number;
  status: string;
  createdAt: string;
}

export interface EnrollmentInput {
  studentId?: number;
  classId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  studentType: string;
}

export interface EnrollmentUpdate {
  status?: string;
}

export interface EnrollmentWithClass {
  id: number;
  studentId: number;
  classId: number;
  status: string;
  createdAt: string;
  class: DanceClass;
}

export interface EnrollmentWithDetails {
  id: number;
  status: string;
  createdAt: string;
  student: Student;
  class: DanceClass;
}

export interface Announcement {
  id: number;
  title: string;
  message: string;
  publishedAt: string;
}

export interface AnnouncementInput {
  title: string;
  message: string;
}

export interface Video {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  classId: number | null;
  /** @nullable */
  className: string | null;
  createdAt: string;
}

export interface VideoInput {
  title: string;
  url: string;
  description?: string;
  classId?: number;
}

export interface DashboardSummary {
  totalStudents: number;
  activeClasses: number;
  totalEnrollments: number;
  announcementsThisMonth: number;
  newStudentsThisMonth: number;
  classesWithWaitlist: number;
}

export interface ScheduleEntry {
  day: string;
  className: string;
  time: string;
  location: string;
  instructor: string;
  enrolledCount: number;
  capacity: number;
}

export type ListClassesParams = {
category?: string;
};

export type ListVideosParams = {
classId?: number;
};