# lib/api-zod/src/generated/types/danceClassInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface DanceClassInput {
  name: string;
  style: string;
  description: string;
  instructor: string;
  location: string;
  scheduleDay: string;
  scheduleTime: string;
  price: number;
  pricePeriod: string;
  duration: string;
  ageGroup: string;
  category: string;
  capacity: number;
  spotsAvailable: number;
  colorScheme: string;
}