# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalStudents: number;
  activeClasses: number;
  totalEnrollments: number;
  announcementsThisMonth: number;
  newStudentsThisMonth: number;
  classesWithWaitlist: number;
}