# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Enrollment {
  id: number;
  studentId: number;
  classId: number;
  status: string;
  createdAt: string;
}