# lib/api-zod/src/generated/types/enrollmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface EnrollmentInput {
  studentId?: number;
  classId: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  studentType: string;
}