# lib/api-zod/src/generated/types/enrollmentWithClass.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { DanceClass } from './danceClass';

export interface EnrollmentWithClass {
  id: number;
  studentId: number;
  classId: number;
  status: string;
  createdAt: string;
  class: DanceClass;
}