# lib/api-zod/src/generated/types/scheduleEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleEntry {
  day: string;
  className: string;
  time: string;
  location: string;
  instructor: string;
  enrolledCount: number;
  capacity: number;
}