# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Happy Feet Dance Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  studentType: string;
  createdAt: string;
}